backend/requirements.txt
flask
flask-cors
requests
firebase-admin
python-dotenv
pyDolarVenezuela
mysql-connector-python